refactor(batcher): simplify L2 block sourcing to polling-only
