+++ crates/batcher/service/src/config.rs
@@ -36,12 +36,6 @@ pub struct BatcherConfig {
     /// [`l1_rpc_url`](Self::l1_rpc_url) only on failure. When absent, polling
     /// is used exclusively.
     pub l1_ws_url: Option<Url>,
-    /// Optional L2 WebSocket endpoint for new-block subscriptions.
-    ///
-    /// When set, the batcher subscribes to new block headers over this
-    /// connection and falls back to polling [`l2_rpc_url`](Self::l2_rpc_url)
-    /// only on failure. When absent, the batcher uses polling exclusively.
-    pub l2_ws_url: Option<Url>,
     /// Optional derived-parity validator L2 RPC endpoint.
     ///
     /// When set, shadow-mode deployments compare derived L2 block hashes
@@ -131,7 +125,6 @@ impl Default for BatcherConfig {
             l1_rpc_url: vec!["http://localhost:8545".parse().expect("valid default URL")],
             l1_ws_url: None,
             l2_rpc_url: vec!["http://localhost:9545".parse().expect("valid default URL")],
-            l2_ws_url: None,
             parity_validator_l2_rpc_url: None,
             rollup_rpc_url: vec!["http://localhost:7545".parse().expect("valid default URL")],
             l1_beacon_url: None,
