+++ crates/batcher/source/README.md
@@ -2,14 +2,13 @@
 
 L2 unsafe block source for the Base batcher.
 
-Provides the `UnsafeBlockSource` trait and `HybridBlockSource` implementation
-that combines WebSocket subscription and HTTP polling with deduplication,
-sequential gap recovery, and reorg detection.
+Provides ordered L2 block polling, reorg detection, and L1 head tracking.
 
 ## Components
 
 - **`UnsafeBlockSource`** — async trait for streaming L2 block events
 - **`L2BlockEvent`** — new block or reorg signal
-- **`HybridBlockSource`** — merges live updates with ordered polling
+- **`PollingBlockSource`** — fetches consecutive L2 blocks above a safe head
 - **`PollingSource`** — trait for fetching an L2 block by number
+- **`HybridL1HeadSource`** — combines L1 subscription and polling
 - **`InMemoryBlockSource`** (`test_utils`) — in-memory source for action tests
